requirements.txt
requests
python-crontab
pycurl